Rebuild the param building logic to not assume anything
I quickly ran into a case where the names of the keys were just wrong
and had to rethink this. I've now set uppon requiring that the hash keys
are the names of the parameters in the AWS API, which I believe makes this
easier to read, use, and debug.