Improve handling of S3 post body sizes and content-types. Fixes #11
Added a check to see what the given body responds to when checking the
content size. This will let users upload raw body content instead of
using a File. From this I found out that Net::HTTP will default the
content-type of POST requests to x-www-form-urlencoded so we need to
handle this case ourself or Amazon's signing craps out.